pyproject.toml
[project]
name = "service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.73",
    "dvc[gdrive]>=3.66.1",
    "httpx[http2]>=0.28.1",
    "pandas>=3.0.1",
    "playwright>=1.58.0",
    "playwright-stealth>=2.0.2",
    "pyarrow>=23.0.1",
    "python-dotenv>=1.2.2",
    "tqdm>=4.67.3",
]
